forgot to address changed in codes of exam 2
new fucntion MatLoad(e) which will take link to the texture and return a material with some of its other map(which is highly un-reusable). Will find a way to load a whole folder soon. Also, we are going to reach 1000 lines.  Maybe i should find a way to seperate files